backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_resources==6.4.4
iniconfig==2.0.0
jiter==0.5.0
limits==3.13.0
numpy==2.1.0
openai==1.43.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
